:rocket: Split models between its apps
