+++ app/main/__init__.py
@@ -3,7 +3,7 @@ from flask import Blueprint
 main = Blueprint('main', __name__)
 
 from . import views, errors
-from app.models import Permission
+from app.auth.models import Permission
 
 
 @main.app_context_processor
